feat: merge and enhance 3rd player void test scenarios
Consolidate thirdPlayerVoidIssue.test.ts into thirdPlayerVoidLogic.test.ts
with comprehensive test coverage and improved memory mocking infrastructure.

Key improvements:
- Merged 6 test scenarios covering all 3rd player void logic
- Enhanced memory mocking helpers in test infrastructure
- Strategic trump expectations (Queen vs low trump) for forcing 4th player
- Fixed TypeScript and ESLint issues
- Comprehensive coverage: 2nd player trumped, both players void, legacy scenarios
